third_party/WebKit/LayoutTests/virtual/feature-policy/http/tests/feature-policy/vibrate-disabled-expected.txt

CONSOLE ERROR: line 9: Navigator.vibrate() is not enabled in feature policy for this frame.
 

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
This is a testharness.js-based test.
PASS No iframe may call navigator.vibrate when disabled. 
Harness: the test ran to completion.


--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
This is a testharness.js-based test.
PASS No iframe may call navigator.vibrate when disabled. 
Harness: the test ran to completion.
